client: rewrite session api
This patch proposes a modified session api with the aim of being simpler
and more robust to use.

File-based and memory-based transfers are clearly separated, and the
functions have been renamed to avoid any confusion.

Even though the new functions are similar to the previous ones, note
that there are important changes in signature and parameter order. For
example, obc_session_put_file takes the remote name before the local
filename, as opposed to obc_session_send in the previous api.

Transfer callbacks receive a pointer to the transfer, and thus the
function obc_session_get_transfer is not needed any more.

obc_session_register has also been removed because the D-Bus registation
of the session is handled internally.

After this change, both put and get operations will be queued equally.